Add game scene for chapter 2 and adjust configs
Move old game scenes under `lib/scenes/legacy`
